Add delete method, enum the S3ResourceTool
Added delete to the crud methods on the BaseResource object. Users can now delete the firebase reference to their resources.
Implementations will have to handle deleting the underlying external resource for now.

In my first PR I missed a string compare type check on "vortex" tool name, so I converted the S3ResourceTool type from an Interface to an enum, and added a general runtime check. To add new tools we only need to add a new enrtry to the enum.  As Scott said, we probably want to move all of that to
Part of PT Story "Store Templates on S3" for the Vortex data collection tool.

[#171094550]
https://www.pivotaltracker.com/n/projects/736901/stories/171094550